Functioning log out and log in

+++ TravelUP/src/main/java/org/greencoding/TravelUP/controllers/AuthenticationController.java
@@ -32,7 +32,7 @@ public class AuthenticationController {
 
     public User getUserFromSession(HttpSession session) {
 
-        Integer userId = (Integer) session.getAttribute(userSessionKey);
+        Long userId = (Long) session.getAttribute(userSessionKey);
         if (userId == null) {
             return null;
